Advanced search modification for Technical View component
